// src/renderer/src/shared/store/useConfigStore.d.ts
export type TConfigDataProps = {
  kick: {
    slug: string
    id?: number
    user_id?: number
  }
  twitch: {
    channel: string
  }
  youtube: {
    channelName: string
    channelId?: string
  }
  x: number
  y: number
  width: number
  height: number
  font: {
    size: number
    weight: number
  }
  messageDuration?: number
  alertSound?: 'none' | 'alert1' | 'alert2'
}
type HideWindowProps = {
  config: TConfigDataProps | null
  setConfig: (data: HideWindowProps['config']) => void
}
export declare const useConfigStore: import('zustand').UseBoundStore<
  import('zustand').StoreApi<HideWindowProps>
>
export {}