// src/renderer/src/shared/constants/defaultConfig.d.ts
export declare const DEFAULT_CONFIG_DATA: {
  kick: {
    slug: string
  }
  twitch: {
    channel: string
  }
  youtube: {
    channelName: string
  }
  font: {
    size: number
    weight: number
  }
  messageDuration: number
  alertSound: 'none'
  x: number
  y: number
  width: number
  height: number
}